break out the code that computes VerifyBounds
Later, we'll defer this work until a separate phase.
