+++ src/librustc/infer/outlives/mod.rs
@@ -13,3 +13,4 @@
 pub mod env;
 pub mod free_region_map;
 pub mod obligations;
+pub mod verify;
